requirements.txt
cachetools
Appium-Python-Client~=4.4.0
openai~=1.59.7
selenium~=4.27.1
streamlit~=1.41.1
python-dotenv~=1.0.1